refactor(niri): give the projection and pattern decision their own modules
`lattice` answers what an expression denotes and takes `&self` throughout;
`pattern` answers whether a value matches, and its third answer — cannot tell —
is what keeps an undecided match and its no-match trap alive.

A third loose doc turned up on the way: `operand_to_lattice_a`'s, sitting under
the tail of a description of something else entirely. It says what the function
does now, above the function.
